Update text colors for plots
The function besttextcolor in the plotting
module determines the best text color for
the annotations in our heatmaps, given
the value of the cell and the maximum/
mininum values. However, depending on the
colormap, it may be necessary to use different
conditions for setting the text color.

This commit does so by keying the conditions
for the text color to specific colormap names.
We provide support for the inferno, plasma,
and viridis colormaps, as well as the PuRd
colormap. All other colormap names use a default
text annotation condition.

In the future, it might be useful to figure
out if there is some way to make this function
extendable, in the sense of not having to use
the colormap name, but some other property,
such as whether it is diverging or sequential.
Right now, matplotlib colormap objects are not
aware of this particular property they have.
However, even if we *did* have access to such a property,
we would immediately run into the fact matplotlib supports
**reversing** colormaps (for instance, PuRd_r), which
doesn't change the kind of colormap it is, but *does*
change the colors.