Updates to new gate set parameterization/interface.
Moves GateSetParameterization logic back into GateSet class and
separated out calculation logic into its own GateSetCalculator
class, which *could* in the future be used by another GateSet-like
class.

Converted Gate and SPAMVec objects to be numpy.ndarray subclasses,
and re-structured GateSet to *not* be an OrderedDict subclass but
instead contain instances of OrderedDict subclasses ("label dicts"):
rhoVecs, EVecs, and gates.  The idea behind this changes is to
provide the user with a GateSet inteface that can be manipulated
as a dictionary of numpy arrays, but at the same time keeps SPAM
vectors and gates separate and knows how to parameterize itself.