Updated DataSet and MultiDataSet classes to have nicer interfaces.
Added ability to set data within a DataSet using intuitive gate
string and spam label indexing (and for adding data to non-static
DataSets).  DataSet and MultiDataSet classes has been converted
to hold GateString objects instead of just tuples, enabling the
writing of DataSets and MultiDataSets to file without a list of
GateString objects.  This unfortunately prompted a swapping of
argument ordering in the write_dataset and write_multidataset
methods to allow the list of gate strings to be an optional
argument.