Another round of updates to Results object adding "tips" option.
Cleanup of Results object code and templates and addition of
tooltips option for full and brief reports, which enables markup
text and tool tips to assist in mapping report tables and figures
to members within a Results object.  This functionality is enabled
via the boolean "tips" argument to the create_full_report_pdf and
create_brief_report_pdf functions.

Additionally, the Results object now stores (preliminary) metadata
within the produced PDF report files.  This needs to be sanitized
better and possibly added to.

Tests still need to be updated and run, which is probably best to
do after cleaning up the metadata.