Updates to tests: skip tests which require optional imports.
Logic has been added so that tests requiring Image, ImageChops,
deap, and pptx are skipped when the required module cannot be
imported.  A warning message is printed but the tests will no
longer fail.  This is to make the tests more friendly to systems
that don't have these modules, since they aren't strictly
necessary for the bulk of pyGSTi operations.