Fix error raised when making blank boxplot
When making a blank boxplot, the number of non-NaN
boxes is actually 0. This caused errors when we
attempted to set the linlog_trans variable, since
there's a factor of 1/(number of boxes) that comes
into play. We handle this error by adding an if/then
statement to the code which sets linlog_trans to some
(unimportant) value when the number of non-NaN
boxes is 0.