Create StdColormapFactory class
This class serves as a tool by which we can create
the appropriate colormaps and norms for the plots
we are currently making in GST. In particular,
it is capable of returning a canonical Linear/Logarithmic,
Sequential, or Diverging colormaps, and Linear/Logarithmic
or Linear norms.

The boxplot() functions will learn about this class, and
pass instances of it around, in future commits.