Update generate_boxplot and plotting.py functions that call it.
In order to use the linlog colormap we need to know the number of
boxes used in the colormap.  Previously, generate_boxplot took a
matrix-generating function as an argument and generate_boxplot used
this function to create the data to plot.  Now, this plot data is
created in the higher level functions which call generate_boxplot
and create the colormap, so that the number of boxes is available
at the time of colormap creation.

As a side effect, the "interaction" mode and "zoomed" box plots
have been removed from plotting.py, as these types of plots
have not been useful (or used) and incorporating the new colormap
logic into these areas would be unduely burdensome.

Tutorial 05 has been updated and runs, though there are warnings
that need to be addressed, as well as proper behavior of the
linlog colormap when sumUp == True.  Report generation and
test harnesses will also need updating to be compatible with
these changes.