Add "dof" keyword to the StdColormapFactory class
To handle the cases where we sum up individual chi2 boxes
to get an aggregate chi2 value, the number of degrees of freedom
of the chi2 function that data should be distributed as changes.
In particular, if D boxes are summed up, since each box is a chi2_1
random variable, the sum is chi2_D. Hence the need for a changeable
value for the degrees of freedom.

This commit adds the dof keyword argument to the StdColormapFactory
class, with a default of 1, and updates the class' code to ensure
the self.dof value is used when computing an appropriate value
for linlog_trans.