Fix bug in pickling of Results objects & add pretty table printing.
Prior to this fix, Results objects could be pickled and unpickled,
but .tables and .figures would be empty after unpickling (due to
incomplete initialization of the ResultsCache objects).

This fix also adds a __str__ method to ReportTable which allows
for pretty-printing of tables, making it easier for the user to
see what is contained in a ReportTable object.