Add clarifying square roots to fidelity in reports
The fidelity, as defined in the report .tex files, had
a slightly confusing square root - it did not extend
all the way over the process matrix whose square root
was being computed. This looked a bit odd, and could
be fixed by using \sqrt{X} instead of \sqrt(X). This
issue showed up in both the logL and chi2 reports, so
we make the change in both.