Updates and additions to unit tests.  All tests pass, 97% coverage.
Unit tests added to cover recent updates, including:
 - "nested" gatestrings option
 - new GateSet, Gate, and DataSet interfaces
 - new Results interface (ReportFigure, ReportTable objects)
 - "tips" option for reports.
 - non-Markovian error bars.
These tests bring coverage up to 97%, and changes to pygsti code
consists of bug fixes and changes to allow greater testing (e.g.
replacing repeated "raise ValueError" with assert statements). One
notable bug fix was to the copy routines in labelsdict.py, which
caused the label dictionary parent to be inherited instead of being
copied properly in a GateSet.copy() operation.  This caused references
to the parent.povm_identity to be incorrect, thereby making the SPAM-
parameter reported for the "remainder" effect vector to be incorrect
in reports.  Thus, the only update to the accepted "good" reports is
the one spam-parameter number corresponding to the remainder POVM
effect.

It's worthwhile to note here, that even for TP-constrained MLGST, the
resulting GateSet estimate may *not* have a povm_identity that is the
same as the "accepted" target POVM identity -- and because of this the
spam-parameters for E and "1"-E may not add exactly to 1.0.  One could
gauge-optimize the result so that the povm_identity does match the
target identity, but this is currently only one term of many that enter
the gauge optimization objective function.  I think the appropriate
way to handle this is to set the accepted identity upon contracting the
starting seed gateset to TP (this would be considered an aspect of a
gateset being "TP"), so that a later "TP" gauge optimization will not
change the identity.  This is currently *not* implemented yet, as it
will likely result in changes to the "good" reports that should be
 looked at separately and tracked as a distinct commit.