Resolve conflict from previous commit
There were several conflicts in the Analysis tutorial in
the previous merge, which we ignored and committed. As a
result, starting that notebook would cause the server to
crash, since the notebook was not the correct JSON object.
This commit fixes that by adding in the Analysis notebook
from the develop branch. (Essentially, using a "theirs"
strategy for the merge.)