Create new MidPointNorm class
For our diverging colormaps, it is helpful to specifically
set the midpoint of the colormap at 0. This commit teaches
the plotting module a new normalization class, called
MidPointNorm. This norm will take as input a midpoint,
and create a matplotlib Normalization class instance
where the midpoint is mapped to .5 in the norm.

This new norm is useful when we use the "bwr" diverging
colormap, which goes from blue -> white -> red. With
the new norm, we can ensure the midpoint of our data
(which we often want to be 0) is mapped to white, and
low (high) values are mapped towards blue (red).