Teach the StdColorMapFactory class about the new norm
This commit introduces the MidPointNorm class to the
StdColorMapFactory class. We add in a new keyword
argument to the class call, midpoint, which is by
default 0. In addition, when deciding what the
correct normalization to use is, if the colormap
kind is "div", then we use the MidPointNorm class.
(Previously, we were using the standard matplotlib
Normalization class.)