Add 'linlg_pcntle' argument to report template files
This commit incorporates the 'linlg_pnctle' argument into
the GST report templates. It also teaches this argument
to the qtys dictionary within the Results object (which
is used to insert the value into the template), as well
as adds this argument to the default self.parameters()
of the Results object. NOTE: The value here is set as
an integer, and will be divided by 100 when passed to
the plotting functions.

To ensure these values stick out when the reports are compiled
natively in LaTeX, the default value to be inserted is set to
42.