Modify plotting function calls from Results object
Since we have changed some of the plotting functions
(the chi2/logl boxplots, their "direct" versions, and
the whack-a-mole plots) in light of using new colormaps,
some of the arguments passed to these functions from
the Results object are no longer necessary. (For instance,
the vmin/vmax parameters in the normalization.) This
commit removes those parameters from the function calls.

In addition, this commit introduces one stylistic change -
for plots whose x-axis is "L", we use the LaTeX version
r"$L$", which looks a little bit nicer, and is consistent
with the "L" in the report text proper.