Add linlg_pcntle to function calls inside Results object
While previous commits taught the Results object about
the linear/logarithmic transition percentile, and enable
it to include that value inside the GST reports, we had
yet to show it how to actually pass that argument to the
appropriate plotting function! This commit does so by adding
an additional keyword argument inside the chi2/logl boxplot/
direct/whack-a-mole function calls.