Continues MPI testing: removed debug checks & added timers.
Moved debug (tagged w/ "DEBUG!!!") checks to being comments
in testMPI.py, to facilitate making legit unittests in the future.

Initial tests of MPI revealed a very marginal speedup when moving
from 1 to 4 processors (10-20% instead of ~400%), so timers were
added to investigate (tagged with "TIMER!!!", and now commented
out so they could be re-enabled easily).  Timing info showed
that due to:
1) numpy's "dot" call not scaling linearly with the input size
2) MPI overhead
3) the fact that distributing over derivative cols only makes
  the derivative calculation faster (nothing else).
The speed up for 1Q-GST ends up being only 10-20%.  Switching over
to tree-distribution mode shows a better speedup since it affects
everything and scales closer to linear with the tree size.  To this
end, the EvalTree balacing within split(...) was improved to reduce
the average tree size.  Note that there is still no API way to
switch the type of distribution between by-deriv-col and by-tree.