Create a new function for computing number of boxes and DOF
In order to debug why we are having issues generating the pixel
plot appendices (because the code returns the number of boxes as
0, when they clearly are not), it would be helpful to have an
independent function which, when given appropriate GST-related
input, returns the number of boxes, and the number of degrees
of freedom for each box. (This is particularly useful for
making the chi2 and logl boxplots.)

This commit does so, and simply spins off the part of the
_computeSubMxs() function which deal with the number of
boxes and DOF into a new function, _compute_num_boxes_dof().