Change n_boxes calculation when sumUp is False
After some investigation, it was determined the
code for calculating the number of boxes in the chi2/logl
boxplots when sumUp was False was incorrect. Instead of
counting the number of non-NaN boxes, we were counting
the number of NaN boxes. In turn, this cause the code
to crash when making the pixelPlot appendices, since,
for the first GST iteration plot, there were no NaN
boxes.

This commit uses the ~np.isnan() function to compute
whether each element in subMxs is NaN, and then does a
sum to get the total number of boxes.