Adds plumbing, testing, and bug fixes to split-tree Hessian calc.
Previous commits left the ability to compute hessians with a
memory limit mostly implemented but untested.  This commit propagates
a "memLimit" argument up from tools' chi2 and logl_hessian functions,
through get_[chi2|logL]_confidence_region, and finally to a new
parameter (named "memLimit") in the Results class.  This can be set
directly, as is done in the added test (in testReport.py), or via
the "memoryLimitInBytes" advanced option to the do_long_sequence_gst
driver.  Added tests for chi2 and logl hessian computation pass.