Add new functions for computing number of boxes, dof per box
This commit re-works the _compute_num_boxes_dof() function
to utilize more vectorized operations. In addition, it adds
two new functions, _all_same(), and _num_non_nan(), which
are necessary to compute whether all the elements in an array
are the same, and how many non-NaN elements are in an array,
respectively.

The _all_same() function is useful when we use sumUp=True in
some boxplots, because the number of degrees of freedom should
be the same for each box. I add an explicit check as to whether
this is the case, and raise an assert error if they are not.

The _num_non_nan() function is a small helper function which
computes the number of non-NaN elements in an array.