Remove nested array structure for direct_deviation_boxplot
The direct deviation boxplot code was giving me fits with
respect to the _compute_num_boxes_dof() function. I think
the reason was the list of arrays being passed to the function
was excessively nested ([[X]] instead of X). I removed the
additional nesting, and explicitly declared the datatype
of the array as 'float64', replacing the (equally valid,
but a bit more cryptic) 'd' declaration.