Initial work on Hessian computation for 2Q-GST.  So far, takes too long.
This commit adds some needed framework for computing the Hessian in
2Q-GST without running out of memory.  Namely:
 - EvalTree is given a "get_min_tree_size" function for use in
   memory estimates.
 - GateSet (and gscalc) hessian methods (bulk_hpr, bulk_hprods,
   bulk_fill_hprods) now accept a "wrtFilter" argument allowing
   them to compute only a specified set of Hessian columns.  This
   is then used for imposing memory limits, and later may be useful
   for MPI distribution.
 - the logl_hessian function (in likelihoodfns.py) has some
   experimental methods for dividing up the task of computing the
   Hessian when memory is limited.
 - Timing and memory debug printing code in gscalc.py and
   log_lhessian has been commented but not removed, as the work of
   figuring out if & how to compute the Hessian in a reasonable
   amount of time is ongoing.
 - initial version of updated bulk_XXX computation routines
   which should have better code reuse and be slightly faster are
   added to gscalc.py, but not tested yet.  The plan is to have
   these functions replace the existing ones once they've been
   finished and tested.  Currently the only usage of them is
   that the bulk_fill_hprods_new function is now used to compute
   the log-likelihood Hessian (in logl_hessian).

Currenlty, in an 2Q-LSGST, maxL=8, test case, the Hessian
computation takes prohibitively long to compute (100s of hours),
and this appears to be unavoidable so long as we assume we need to
compute a single dot product per gate string.  Time scaling is
as we expect: ~ nParams x (time required for Jacobian), and for
this modest 2Q-GST example nParams=1200 and it takes roughtly 0.5
hours to compute a Jacobian (also shown as being an expected time
given numpy's efficient dot function and a ~1GHz CPU: 0.03s per dot
product x 27000 gate strings ~= 0.5 hours).

All tests pass except reports, which differ only in the updated
text.