Compares new bulk_XXX functions to old. Fixes new so now they agree.
A new, temporary, test file is added to compare the results of the
existing "old" vs. the recently added "new" bulk_XXX functions
(suffixed with _new) within GateSetCalculator.  This comparison
revealed some minor bugs in the "new" methods, which have been
fixed.  Now old and new results agree (to machine precision) and
the next step will be to replace the old routines with the new
ones and remove the temporary test file.

The only changes in function signatures between the old and new
methods is the removal of the "memlimit" argument in bulk_dproduct,
bulk_dpr, bulk_dprobs, and bulk_fill_dprobs.  This argument didn't
actually do anything and was only present to mark the intended
addition of functionality.  However, the current paradigm is to perform
memory management at a higher level and pass *split* evaluation
trees to GateSetCalculator functions to lessen memory consumption,
so there is no need for the argument.