Do gentle text editing of tutorials
This commit fixes some typos and grammatical
errors in the tutorials, as well as adds some
clarifying text in parts. In particular, I tried
to provide a short (one or two sentence) of the
*problem* the tutorial tries to address at the
top.

On the whole, I tried to tackle adding text only
in places where I (think!) I understand enough
about pygsti/GST to say something intelligent.