Perform gentle text editing to tutorial 10
Some basic changes. Generally speaking, I tried to:

- Turn cells that were only comments into markdown cells
- Add in some more helpful comments in the code
- Wrap function calls to ensure they stay within the
width of the cell

The substantial change I made was to re-org the description
of the arguments to optimize_integer_fiducials_slack. I went
with a bulleted list, with the arguments first, and then
a description.

I also used the .format() method on some of the strings
to round the execution time counts.