Fixes bug in non-Markovian error bar generation.
When non-Markovian error bars are requested in a report when the
data was fit very well (and normal error bars *should* have been
used) -- in particular when the actual chi2 (or 2*dlogL) is LESS
than the expected value -- the radius associated with the
non-Markovian error bars can swing negative, resulting a lots of
NANs propagating through the reports (and being shown as dashes).
This is undesirable, and so now a minimum radius of 1e-8 (arbitrary)
is imposed.  In the future, we may want to have the non-Markovian
error bars revert back to the normal ones, but this could be
confusing to the user without some kind of notification.