Adds storage of basis information to GateSet.
The GateSet object is now able to hold a basis name (an abbreviation
like "pp" or "gm") and a basis dimension (a single integer or list
of integers specifying the density-matrix space dimension and structure).
These properties are accessed via the new methods get_basis_name(),
get_basis_dimension(), and set_basis().  This addition is motivated by
our often interpreting certain gate sets as being in a particular
basis, and a desire to "tag" a GateSet with this information.  In
particular, this will facilitate plots/tables where the rows and columns
of gate matrices and SPAM vectors labeled, and the new function
basis_element_labels(...) was created for this purpose.

Currently, basis information is COPIED when GateSets are copied,
SET when build_gateset(...) creates a GateSet OR when a GateSet
is gauge-optimized to a target GateSet that contains basis information,
and is FORGOTTEN whenever a GateSet is loaded from a vector.

This removes the need to specify a "matrix basis" (often mxBasis
in the code) to table generation, plotting, and driver routines that
previously required them, and thereby removes the need for the
Results object to store an 'mxBasis' parameter.