Added tpdist() function to GateSet, fixed parameterization bug.
The function do_long_sequence_gst would fail when the target
gateset was not fully parameterized (i.e. a TP-parameterized gateset).
Initially I thought we should test for TP-ness of a gateset, and so
the tpdist() function was addd to GateSet for this purpose.  However,
it seems better to just full-parameterize a gateset used in
do_long_sequence_gst before gauge optimization when constrainToTP
is True, since we set the parameterization to TP at the end anyway.