WIP for test automation
- Adds python equivalents for most of the testing automation.
- Moves test automation to a new directory called test_tools
- Creates benchmarking package and applies it to some test tools

The goal of this is to make automated tests a little more flexible, and make benchmarking easier.

Right now, a lot of things on this branch don't work correctly, or I can't confirm that they do because I lack the nose library
However, runAllTests and runModule seem to be working.