Modified runPackage.py to fix error when no post command is given.
When running runPackage.py and no postcommand was given, the call
to subprocess would pass an empty string as the first argument.
This caused unittest.main to complain that it couldn't find the
module ''.  This was fixed by restricting the arguments of
unittest.main to only the first passed-in argurment (the executable
name) but this meant that you couldn't run individual tests easily
from the command line.  This fix allows us to pass all arguments
to unittests.main.