Fixes bug in Results object; updates default 'hessianProjection' param.
Adds proper initialization of the _comm member of a Results object,
so that when accessing data within the object *before* creating any
reports you don't get a "_comm is not a member of Results" error.

Also updates the 'hessianProjection' parameter from 'std' to
'optimal gate CIs'.  While this latter option causes confidence
region creation to take more time, it's almost always a more
correct mode to use than 'std' and should thus be the default.