Fixes a few minor syntax errors.  Updates testCore.py.
Adds a missing "_objs." before a reference to VerbosityPrinter
and corrects a misspelling "end_progresss" => "end_progress".

Tests in testCore.py which compare computed to reference
gatesets have been updated to first gauge-optimize the computed
gateset to the reference and then requires only 5 places of
accuracy/precision in the Frobenius norm.  This results in all
the testCore.py tests passing on Erik's desktop, and hopefully on
all other machines as well.