Updates tests so all tests pass on Erik's desktop.
Several tests are updated so that all tests pass on a second
non-Erik's-laptop machine.  The hope is that these modifications
will also result in the tests passing on *any* machine, i.e., the
Erik's-laptop-specific assert statements have been removed.  The
relevant tests now check for the PYGST_DEEP_TESTING environment
variable, and if it is set to ["true"|"1"|"yes"], then these
additional assert statements are enabled (so you should only set
this variable if you happen to be running the tests on Erik's
laptop, or more generally whichever machine is setup to perform
these more detailed tests).  Particular changes are:

- comparison of generated datasets that include randomness is
  only performed when PYGSTI_DEEP_TESTING == true, since different
  versions & installs of numpy can produce different random
  numbers (even when given identical seeds).

- comparison of generated and "reference" LaTeX report files is
  only performed when PYGSI_DEEP_TESTING == true, since numerical
  precision differences result in different values being printed
  on different machines (maybe this can be fixed in the future
  by more intelligent selection of the precision).

- comparison of Hessian eigenvalues to the saved reference values
  is improved to be more robust.