Fixes error with caching non-Markovian ConfidenceRegion objects.
Negative values of the confidence level passed to reporting
functions causes GST to generate "non-Markovian error bars".
A bug caused the so-generated ConfidenceRegion objects to be
incorrectly stored under the negated (hence positive) value
of the confidence level and therefore *not* be used by later
requests for the same non-Markovian confidence region.