Updates reference report tex files.  All tests pass.
Reference report files are updated match the now
consistent output precision.  Previously, reports
had a mix of precision=6 and precision=4 (for vectors
and matrices) and now the precision is set by options
within a Results instance (and is set == 4 in the
tests).  Thus, the only changes to these reference
report files are decreasing precision=6 entries to
precision=4 entries, and this has been verified
my manual examination of the diff output.

All tests, including DEEP testing (on Erik's laptop),
now pass.