Removes gatestring filter from direct-X plots in reports.
Since direct-LGST and direct-LSGST plots use a "sigma" label
the gate strings they use won't match any of the actual gate
sequences used by GST, and so if they are fed the same
gatestring_filter as the usual chi2 or logL box plots they
won't display anything (this is why we needed to handle the
display of all-masked out matrices recently).  So now the
Results object does not pass any gatestring_filter to the
direct-X figures.