Fixes an bug in labeldicts.py (when a duplicate value is set).
OrderedSPAMLabelDict does not allow duplicate values, and
if a value is inserted that already exists, the old key-
value pair is deleted from the dictionary.  This logic
modified the dictionary within a loop iterating over its
elements, which Python3 doesn't like (which is reasonable).
Now the logic is reworked so it doesn't delete any dictionary
elements within the loop iterating over the elements.