Updates linlog color map to work with random reduction reports.
Previously, the number of degrees of freedom (i.e. boxes) within
in each "block" of a color box plot needed to be the same across
all blocks (which results in a well-defined degrees-of-freedom
for defining chi2 distributions in the LinLogNorm color map). This
is not the case when random fiducial pair reduction is used, and
so now a just a warning is given and the *average* number of degrees
of freedom (avg over all the blocks) in a color box plot is used,
which seems fine for now.  Note that this only affects color box plots
when sumUp == True.