Correct GRASP iteration recovery and return value
GRASP will attempt to retry an iteration that raises an exception.
Prior to this change, it would retry an unbounded number of times,
which could bring your system to a crawl if some exception was
repeatedly being raised very quickly. Now, if an iteration raises
an exception 11 times in sequence, GRASP will stop trying to
suppress the exception and raise it itself.

Also corrects the return if else logic so that it operates on the
entire tuple to be returned in the "if" case instead of only the
last element of that tuple.