Include tools for evaluating germ set performance
Adds an function that facilitates simulating GST experiments with
a given germ set under a variety of conditions (gatesets of varying
distance from the target, different max sequence lengths, different
numbers of measurements taken, etc.). Also adds an object to wrap
the results of said simulations and provide helpful visualizations
for evaluating the germ set's effectiveness.

The function lives in drivers, and the object lives in objects.