Let GRASP shuffle search nghbrhds & show progress
Shuffling search neighborhoods could be important for generating
solution diversity when the local optimizer updates the solution to
the first better solution it finds (rather than exhaustively
searching the neighborhood and taking the absolute best).

Fixes a few unnoticed bugs in germ and fiducial GRASP and allows
them to display iteration number to stdout.