Remove GermSetEval from objects/__init__.py
There is some circular import problem that python 2 complains about
with objects/germseteval.py importing from
algorithms/germselection.py which imports objects. It works fine in
python 3, but it won't run in python 2. Perhaps there needs to be
some refactoring of the germ selection code where the parts that
GermSetEval needs are split off into tools (which hopefully
wouldn't need to import from objects), but for now we'll just
remove the automatic import of germseteval with an objects import,
which seems to make things work.