Correct printer logging.
Many log messages were being logged with no explicit verbosity
level, meaning they would be printed out even if the verbosity of
the printer were explicitly set to 0. This commit sets all logging
in germ and fiducial selection to an explicit verbosity so that
setting the printer verbosity to 0 gives silent output as desired.