Applies minor fixes to figure saving.
Removes code which initializses callbacks for figures and
ultimatey closes figures within the ReportFigure constructor.  We
shouldn't need to set the callbacks in this case, and in plotting.py
it is assumed that the matplotlib figure remains open after creating
a ReportFigure - so we don't want to close the figure in
ReportFigure.__init__.  This may have caused some weird test failures
regarding matplotlib not wanting to plot infinite values (?), but
it's unclear to me if/how these issues are connected.