Makes runTests.py more robust when --cover is used.
Previously, when --cover was used, runTests.py would always call
"coverage combine" after running nose tests.  This worked on some
machines but not others, where "coverage combine" in a directory
with just a single ".coverage" file would overwrite that file with
an empty-coverage one (because it was only looking for
.coverage.* files to combine, and there weren't any).  I'm not sure
why the behavior is different on different machines (with identical
nose and coverage versions!), but now runTests.py only combines
coverage files when run in parallel, in which case .coverage.<pid>
files are always created so there should be no issues.