Updates travis CI deployment - only when all jobs pass.
Removes the built-in Travis-CI PyPi deployment, since this fires
whenever any one of the jobs in a build passes, and adds manual
PyPi uploading to the CI/deploy script (which probably won't work
yet), which is only executed after *all* the jobs pass.

Also adds conditional branch name checking so that merging with
beta is *only* attempted on the develop branch (and PyPi deployment
is only attempted on the master branch).