Fixes a bug in fake data generation.  Changes behavior slightly.
Tests were failing because an all-zero count dict was being
computed by generate_fake_data, which then wasn't being added to
the dataset ultimately causing a KeyError.  This was due to a
bug in the probabilty-adjustment code which used a (len-1) instead
of just (len) to perform an adjustment.  This inspection also revealed
that such adjustment is only necessary when the sample error is
either binomial or multinomial - and so now these adjustments are
make *only* when needed.  Thus, if sample error is set to "round"
then the probabilities can be wacky and generate_fake_data won't
complain (but it won't give counts that add to nSamples either).

This commit also removes previous adjustment code specific to the
"binomial" case, which is no longer necessary.