Update to verbosity printer progress display & tests.
Progress bar (hash marks) now fills in hashes up-to but *not*
including the current iteration, and _end_progress finishes
the bar with one additional iteration (usually making the bar
full).  For example, on iteration 1 of 2 the bar is *empty* now,
not half full; on iteration 2 of 2 the bar is half full, and once
the progress context has ended the bar is completely full.  This
seems to best track with the actual progress, since usually we
print the show_progress method at the *beginning* of the iteration
(and this is how we *should* keep doing it!).

Verbosity printer tests have now been updated due to these changes
and those of the last commit or two, so now printer tests (and I
think all the others too) should pass.