Adds nan handling to two_delta_loglfn
Now the function two_delta_loglfn, used to create logl box  plots,
can deal with having nan elements in the arrays (ptic the f
 argument) passed to it, meaning it doesn't show RuntimeWarnings
anymore (which are annoying and don't actually signify anything
is wrong).  It is normal to have nans in the arguments to this
function since this tells the plotting routines not to plot or
compute those items, e.g. in fiducial pair reduction.