Updates logl_boxplot and chi2_boxplot to avoid unecessary computation.
Previously, logl_boxplot and chi2_boxplot were taking a very long
time to run because they were pre-computing probabilities for
*all* of the gate strings in the given dataset (even though not
all were necessarily used).  This has been fixed now, by pre-
computing all of the "gatestring maps" ahead of time and
concatenating them to form the list of gate sequences needed for
the entire plot.  Debugging this lead to the insertion of timing
traces (print statements), which have been left in but commented
out.  This modification leaves the pre-compute code in
logl_boxplot and chi2_boxplot slightly messy and should probably
be consolidated & cleaned up in the future.