Adds "timer dictionary" plumbing to do weak profiling.
New timer_dict arguments have been added to some core.py,
gateset.py, and gscalc.py functions to allow the accumulation of
run times across multiple calls to the function.  For example,
the total time spend in the jacobian functions can be aggregated
this way.  The "timer" elements of timer_dict are added to the
list of times that eventually get stored in a Results object.

This was implmented as a way of check & debug MPI functionality.