Adds initial implementation of "Intelligent Fiducial Pair Reducion"
Intelligent Fiducial Pair Reduction (IFPR) is an algorithm that
selects a subset of all fiducial pairs on a per-germ basis so as
to ensure that all of the germ's eigenvalues (and certain off-
diagonal elements in the germ's diagonal basis) can be resolved.
This algorithm has similarities with fiducial selection in LGST
in that we look for a "full-rank" projection of a Gram-matrix
Jacobian onto the space of eigenvalues (+certain off-diag els).

Advantages of this procedure when compared with the previous FPR
include speed and a more rigorous mathematical footing.

In the implementation, a key piece is the new
EigenvalueParameterizedGate class, which implements a gate whose
only parameters are its eigenvalues (+certain off-diag els).  The
IFPR algorithm is invoked using the function
find_sufficient_fiducial_pairs_per_germ, which may deserve a better
name in the future.

As a side effect, the fidPairs argument of many of the plotting,
driver, and gatestring-list-generation routines can now be a
dictionary specifying fiducial pairs to keep on a per-germ basis.
This prompted refactoring of plotting.py so that the "fidPairs"
argument has been renamed "fidpair_filter" and serves a more
parallel role to the "gatestring_filter" argument in all functions.

Report generation code has been modified accordingly with the
end result that one should be able to perform end-to-end GST
using IFPR gatestring sets.

A simple unit test has been added to ensure basic functionality of
the code.  No actual performance or algorithm correctness tests
have been done yet - we just know the code runs.