Updates custom-LM routine amidst performance testing.
This commit adds untested "debug-mode" updates to the custom-LM
routine in order to easily test on different HPC systems.  This
commit point is *very* ustable and full of debug statements.
Updates include:
- a tweak to the LM routine which changes the way
the damping is applied (old: multiplier (mu) * Identity, new:
multiplier (mu) * diag(JTJ)) which in the single 2Q-GST case
I've been testing has resulted in a substantial speedup.
- addition of memory printing/profiling stmts to custom_leastsq
- addition of timer profiling to custom_leastsq
- substitution of seral dot(J.T,J) with mpitools.mpido(J.T,J)
  along with some supporting code.  This is completely untested
  (HPC queue too full at this time to even test if it runs).

The central goal in view, for reference, is to find the timing and
memory bottlenecks when running GST on multiple CPUs and eliminate
them.