Adds many debug statements (most commented out now).
This commit marks a peak in debug statements attempting to track
memory allocations/deallocations.  The motiviation was that it
appears that there is some kind of memory leak, as successive
interations consume more and more memory.  However, through the
use of numpy's allocation tracking hooks this doesn't seem to be
occuring (alloc'd mem levels stay constant...) so something is
fishy there.  In the end, the takeaway message seems to be that
numpy cannot be trusted to allocate & deallocate from the heap
efficiently enough to make good memory estimates. For instance,
in some of the commented out debug statements, we can see that
just performing an unrelated numpy.ones(...) allocation before
"freeing" (setting to None) another allocated array causes some
of the latter memory to be held and seemingly never released.

I think the best course forward is to eliminate as much as possible
the copying of large arrays of data.  Even though numpy copies are
fast, they can be the source of leaked/inefficiently-deallocated
blocks of memory which can cause the required memory to rachet up
over time.  The first step in doing this is to update EvalTree so
that subtrees' "final" elements are proper slices of the parent
tree's final elements.  After this, we should be able to place
computed (subtree, column_block) results directly into their
output array (and MPI broadcasted/gathered from there) instead
of copying along intermediate arrays.